requirements.txt
matplotlib
numpy
pillow
scikit-learn
scipy
